handle empty collectibles state
