+++ examples/xnft/iframe/package.json
@@ -3,10 +3,10 @@
   "version": "0.1.0",
   "scripts": {
     "build": "xnft build",
-    "start": "xnft watch"
+    "start": "xnft watch",
+    "dev": "xnft dev"
   },
   "devDependencies": {
-    "@coral-xyz/xnft-cli": "*",
     "@parcel/config-default": "^2.7.0"
   },
   "dependencies": {
